Check for container restarts (#117)
When we run samples using k8s-sandbox, the agent can exceed the memory limits and force the container to restart. There is a section about this in the docs. The agent can't detect the restart, so it gets confused that the sandbox has been reset. (Slack)

This PR makes the framework track the pod UID and the number of times the default container has been restarted.

It then validates these values before performing any operations, and either warns or raises an error if it detects a change.

* Check for container restart
* +doc
* Fix no container when warning
* whitespace
* Add pod name to warning/error message
* Format