Add tests for initContainer
There is some duplication of resources across two tests. They could be
unified, but it might be better to do this in a separate PR and limit
the scope of this PR. A future PR can
look at the `install_sandbox_environments` signature, which can be simplified to
make references to `resources/` files more flexible.