Adding support for initContainers (#135)
* Rejig test invocations

This is to support testing locally

* Add support for initContainers

Useful when a main container’s initialization depends on another pod being ready.
Sometimes the cross-pod dependency requires a separate image from the main container.
Other use cases include situations where the main container’s startup script is complex, making it cleaner to handle the dependency in an initContainer.

* Add tests for initContainer

There is some duplication of resources across two tests. They could be
unified, but it might be better to do this in a separate PR and limit
the scope of this PR. A future PR can
look at the `install_sandbox_environments` signature, which can be simplified to
make references to `resources/` files more flexible.

* Revert "Rejig test invocations"

Reverting based on: https://github.com/UKGovernmentBEIS/inspect_k8s_sandbox/pull/135#issuecomment-3543017161

This reverts commit a0585e3bbf1219cf69abcddbf54ecd30f73ee5de.

* Additional changes

* prepare for next release

* fix changelog

---------

Co-authored-by: Art.OCathain <Art.OCathain@dsit.gov.uk>